added a textView generator
