+++ ReadingList/app/src/main/java/com/thadocizn/readinglist/Constants.java
@@ -1,0 +1,5 @@
+package com.thadocizn.readinglist;
+
+public class Constants {
+    public static final String CSV_STRING = "csv_string";
+}
