+++ ReadingList/app/src/main/java/com/thadocizn/readinglist/EditBookActivity.java
@@ -1,0 +1,13 @@
+package com.thadocizn.readinglist;
+
+import android.support.v7.app.AppCompatActivity;
+import android.os.Bundle;
+
+public class EditBookActivity extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_edit_book);
+    }
+}
