+++ ReadingList/app/src/main/java/com/thadocizn/readinglist/MainActivity.java
@@ -1,0 +1,36 @@
+package com.thadocizn.readinglist;
+
+import android.content.Context;
+import android.content.Intent;
+import android.support.v7.app.AppCompatActivity;
+import android.os.Bundle;
+import android.view.View;
+import android.widget.TextView;
+
+public class MainActivity extends AppCompatActivity {
+
+    Context context;
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_main);
+        context = this;
+    }
+
+    private TextView getTextView(final Book book) {
+        TextView textView = new TextView(this);
+        textView.setText(book.getTitle());
+        textView.setTextSize(24);
+        textView.setPadding(10, 10, 10, 10);
+        textView.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                Intent intent = new Intent(context, EditBookActivity.class);
+                intent.putExtra(Constants.CSV_STRING, book.toCsvString());
+                startActivity(intent);
+            }
+        });
+        return textView;
+    }
+
+}
