+++ ReadingList/app/src/main/res/layout/activity_main.xml
@@ -3,9 +3,9 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
-    android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".activities.MainActivity">
+    android:layout_height="match_parent"
+    tools:context=".MainActivity">
 
     <ScrollView
         android:layout_width="match_parent"
@@ -21,8 +21,9 @@
     </ScrollView>
 
     <Button
-        android:id="@+id/btnAddBook"
-        android:layout_width="match_parent"
+        android:id="@+id/button"
+        android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:text="Add a Book" />
+        android:text="Button" />
+
 </LinearLayout>
