+++ ReadingList/app/src/main/res/values/strings.xml
@@ -1,5 +1,3 @@
 <resources>
-    <string name="app_name">Reading List</string>
-    <string name="textview">TextView</string>
-    <string name="name">Name</string>
+    <string name="app_name">ReadingList</string>
 </resources>
